bugfix: strong params exception when using require
I was using scope: with my form to group the data together and using
params.require(:flight).permit(...)
But when no form submition was made and the url was just '../' instead
of '../..?..' it threw an exception that :flight was empty.
Makes perfect sense when looking back. Spent 4 hours on this cheeky
little thing. Onwards!!